feat: separate create-btn as a component
