+++ src/routes/Home.css
@@ -246,62 +246,3 @@
 .book-list-wrapper:nth-child(1){
     border-top: 1px solid #B6BBC0;
 }
-
-/* create-button-container */
-.create-button-container {
-    position: relative;
-}
-.create-btn {
-    position: fixed;
-    bottom: 60px;
-    right: 60px;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    background-color: #FC7C12;
-    border-radius: 30px;
-    padding: 10px 20px;
-    cursor: pointer;
-    z-index: 100;
-}
-.create-btn img {
-    width: 20px;
-    height: 20px;
-    margin-right: 5px;
-}
-.create-btn span {
-    font-size: 16px;
-    font-weight: 600;
-    color: #fff;
-}
-.create-option-wrapper {
-    list-style: none;
-    border: 1px solid #CCCBC8;
-    border-radius: 20px;
-    padding: 10px;
-    position: fixed;
-    bottom: 95px;
-    right: 60px;
-    z-index: 100;
-    background-color: #fff;
-    box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.25);
-    display: none;
-}
-.create-option-wrapper.show {
-    display: block;
-}
-.create-option-wrapper > li {
-    font-size: 16px;
-    cursor: pointer;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: left;
-    padding: 5px 10px;
-}
-.create-option-wrapper > li > img {
-    width: 20px;
-    height: 20px;
-    margin-right: 10px;
-}
