+++ src/routes/Home.js
@@ -2,14 +2,10 @@ import React, { useState } from "react";
 import { bookList, postList } from "./data";
 import { Link } from 'react-router-dom';
 import './Home.css'
-import BookItem from '../components/BookItem';
+import BookItem from "../components/BookItem";
+import CreateBtn from "../components/CreateBtn";
 
 function Home() {
-    const [createBtnOpen, setCreateBtnOpen] = useState(false);
-
-    const handleCreateBtn = () => {
-      setCreateBtnOpen(!createBtnOpen);
-    }
     return (
       <div className="main-container">
         <div className="left-container">
@@ -83,30 +79,7 @@ function Home() {
               </ul>
             </div>
         </div>
-        <div className="create-button-container">
-            <div className="create-btn" onClick={handleCreateBtn}>
-                <img src={require("../assets/icons/plus.png")} alt="" />
-                <span>등록하기</span>
-            </div>
-            <ul className={`create-option-wrapper ${createBtnOpen ? 'show' : ''}`}>
-                <li className="post-create-btn">
-                    <img src={require("../assets/icons/post-create.png")} alt=""/>
-                    <span>게시글 작성</span>
-                </li>
-                <li className="shre-create-btn">
-                    <img src={require("../assets/icons/share-create.png")} alt=""/>
-                    <span>책 나누기</span>
-                </li>
-                <li className="rent-create-btn">
-                    <img src={require("../assets/icons/rent-create.png")} alt=""/>
-                    <span>책 대여 등록하기</span>
-                </li>
-                <li className="sell-create-btn">
-                    <img src={require("../assets/icons/sell-create.png")} alt=""/>
-                    <span>책 판매하기</span>
-                </li>
-            </ul>
-        </div>
+        <CreateBtn />
     </div>
     );
 }
